Fix helper import
